Oops, improve series name parsing in Styling Studio auto-importer
Ahh right, that's part of why I skipped Prismatics: it's no longer true
that the first word of the style is its series name.

In this change, I try to parse out everything before the pet name part
of the style name, and default to skipping it if we can't quite get it
right.