Run `bundle:update`, and work around bug in latest `protocol-rack`
It's been a while, time to update all our minor/patch versions!

When I first ran this, the dev server stopped working entirely, saying:

```
NoMethodError: undefined method `ip_address' for an instance of Socket
```

I traced this back to what seems to be a bug introduced in
`protocol-rack` version `0.11.0`, so I modified the Gemfile to hold us
back from that version in particular. After that, everything is looking
good!