Fix bug parsing discount time when importing from NC Mall
Previously, we took the incoming date as just an integer, which was
raising a database error when we attempted to save it to a timestamp
field. Now, we correctly parse it into a timestamp.

I imagine we just haven't run the import manually while there's a real
discount in the records before? Weird!